Add files via upload
The objective of this customer churn analysis is to identify and understand the factors that contribute to customer attrition in the credit card business of the North American bank. 
By analyzing the churn data, the analysis aims to uncover patterns, trends, and potential drivers of customer churn. 
The ultimate goal is to provide insights and recommendations to the bank on how to reduce the customer loss rate and improve the performance of its credit card business. 
This analysis will help the bank make informed decisions and implement targeted strategies to retain existing customers and attract new ones.
Dataset Description:
The data set was collected from the North American Bank.
It was having  20 columns, the objective is to reduce the customer loss rate so we are observing the reason for attrition.
To know the reason for attrition I’m considering the columns such as Attrition_flag, Gender, Card_category, Region, clientnum and Income category.
Clientnum: Unique identifier for each customer
Attrition_Flag: Indicates whether the customer has churned or not.
Income_Category: The income category of the customer.
Card_Category: The category of the credit card held by the customer.
Region: The region where the customer is located.