Add files via upload
Retrieve and explore a movie dataset containing relevant information such as movie titles, genres, production budgets, box office revenues, and production companies.

Determine the key factors that contribute to a movie's success, such as genre, production budget, revenue and runtime by analyzing the dataset.
Extract meaningful insights from the analysis and present recommendations to the production company regarding movie genres, production budgets, release dates, or other influential factors to improve their chances of commercial success and positive audience reception.
Dataset Description:
The movie dataset was sourced from TMBD dataset .
It consists of 20 columns , the objective is to predicting movie success and providing insights to the production company for my analysis I’m considering the below columns.
Genre : The genre of the movie and TMDB id (in JSON format) 
Budget : The budget of a movie in dollars.
Title : English title of a movie
 production_company : All production companies' names and TMDB id of a movie (in JSON format)
Revenue : The total revenue earned by a movie (in dollars).
Runtime : The total runtime of a movie in minutes (integer). 
id  :  IMDB id of a movie (string).