Add files via upload
The goal of this project is to study and analyze the right group of new customers for an automobile company, so the company can adopt the specific proven marketing strategy to each of them and be more successful in the business.
The aim of this project is to help Automobile company identify their existing customer trends and behavior to drive more value for their organization. We will be using the existing customer demographic, customer address, and transaction datasets as a labelled dataset to drive the most value customers for the organization.